backend/foodgram-api/requirements.txt
asgiref==3.6.0
Django==3.2
django-cors-headers==4.1.0
django-debug-toolbar==3.2.4
django-filter==22.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.2.2
djoser
flake8==5.0.4
flake8-broken-line==0.6.0
flake8-isort==6.0.0
flake8-plugin-utils==1.3.2
flake8-return==1.2.0
isort
gunicorn
pep8-naming==0.13.3
pigar==2.0.8
psycopg2-binary
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python-dateutil==2.8.2
python-dotenv
Pillow
requests==2.26.0
requests-oauthlib==1.3.1
sorl-thumbnail==12.7.0
sqlparse==0.4.3
